pyproject.toml
[tool.poetry]
name = "lucidtask-backend"
version = "0.1.0"
description = "A lucidtask for upwork job post"
authors = ["Dinesh <dinezh4@gmail.com>"]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.111.0"
sqlalchemy = "^2.0.30"
passlib = {extras = ["bcrypt"], version = "^1.7.4"}
pyjwt = "^2.8.0"
cachetools = "^5.3.3"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
